[ACS-4539] Make ms-office as secondary point of aca-content
