+++ .github/actions/update-library-versions/update-lib-versions.sh
@@ -12,7 +12,6 @@ fi
 export PROJECTS=(
   'aca-content'
   'aca-folder-rules'
-  'adf-office-services-ext'
   'aca-preview'
   'aca-shared'
   'aca-viewer'
