+++ projects/aca-content/ms-office/assets/aos.plugin.json
@@ -1,5 +1,5 @@
 {
-  "$schema": "../../../extension.schema.json",
+  "$schema": "../../../../extension.schema.json",
   "$id": "9a635542-d87a-4558-ae64-ffa199d1a364",
   "$version": "0.0.10",
   "$name": "Edit in Office",
