+++ projects/aca-content/ms-office/assets/i18n/fr.json
