+++ projects/aca-content/ms-office/src/public-api.ts
@@ -22,23 +22,8 @@
  * from Hyland Software. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { ExtensionService, provideExtensionConfig } from '@alfresco/adf-extensions';
-import { NgModule } from '@angular/core';
-import { EffectsModule } from '@ngrx/effects';
-import { AosEffects } from './effects/aos.effects';
-import { TranslationService } from '@alfresco/adf-core';
-import { AlfrescoOfficeExtensionService } from '@alfresco/aca-shared';
-import { canOpenWithOffice } from '@alfresco/aca-shared/rules';
+export * from './aos-extension.service';
+export * from './actions/aos.actions';
+export * from './effects/aos.effects';
 
-@NgModule({
-  imports: [EffectsModule.forFeature([AosEffects])],
-  providers: [provideExtensionConfig(['aos.plugin.json'])]
-})
-export class AosExtensionModule {
-  constructor(extensions: ExtensionService, translation: TranslationService, aosService: AlfrescoOfficeExtensionService) {
-    translation.addTranslationFolder('adf-office-services-ext', 'assets/adf-office-services-ext');
-    extensions.setEvaluators({
-      'aos.canOpenWithOffice': (context) => aosService.isAosPluginEnabled() && canOpenWithOffice(context)
-    });
-  }
-}
+export * from './aos-extension.module';
