+++ projects/aca-content/ng-package.json
@@ -12,6 +12,7 @@
       "output": "ui"
     },
     "about/assets",
+    "ms-office/assets",
     "assets"
   ]
 }
