--- projects/adf-office-services-ext/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@alfresco/adf-office-services-ext",
-  "version": "2.0.1",
-  "license": "LGPL-3.0",
-  "main": "src/index.ts",
-  "dependencies": {
-    "tslib": "^2.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/Alfresco/alfresco-content-app.git"
-  }
-}
