--- projects/adf-office-services-ext/tsconfig.lib.prod.json
@@ -1,14 +1,0 @@
-{
-  "extends": "./tsconfig.lib.json",
-  "compilerOptions": {
-    "declarationMap": false,
-    "paths": {
-      "@alfresco/aca-shared": ["dist/@alfresco/aca-shared"],
-      "@alfresco/aca-shared/store": ["dist/@alfresco/aca-shared/store"],
-      "@alfresco/aca-shared/rules": ["dist/@alfresco/aca-shared/rules"]
-    }
-  },
-  "angularCompilerOptions": {
-    "compilationMode": "partial"
-  }
-}
