--- projects/adf-office-services-ext/tsconfig.spec.adf.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.adf.json",
-  "compilerOptions": {
-    "outDir": "../../out-tsc/spec",
-    "types": ["jasmine", "node"]
-  },
-  "files": ["src/test.ts"],
-  "include": ["**/*.spec.ts", "**/*.d.ts"]
-}
